hw-mgmt: psu fw update tool: rename scripts to be convention like.
Its still not in hw-management convention because according to
 python naming convention "-" can't be used, only "_".
