+++ usr/usr/bin/hw_management_psu_fw_update_common.py
